latest

+++ lib/windot_products/catalog/product_item.ex
@@ -3,7 +3,7 @@ defmodule WindotProducts.Catalog.ProductItem do
   import Ecto.Changeset
 
   schema "product_items" do
-    field(:quantity, :integer)
+    field(:quantity, :float)
 
     belongs_to(:material, WindotProducts.Catalog.Material)
     belongs_to(:product, WindotProducts.Catalog.Product)
